Make HandlerLoader use ScanClassProvicder
