+++ event/core/src/main/java/io/agistep/event/EventHandlerLoader.java
@@ -1,68 +1,34 @@
 package io.agistep.event;
 
 import io.agistep.utils.AnnotationHelper;
+import io.agistep.utils.ScanClassProvider;
 import org.apache.commons.lang3.tuple.Pair;
 
-import java.io.File;
-import java.io.IOException;
 import java.lang.reflect.Method;
-import java.net.URL;
-import java.util.*;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 import static java.util.stream.Collectors.toList;
 
-class EventHandlerLoader {
+public class EventHandlerLoader {
     private static Map<String, HandlerAdapter> handlers = new HashMap<>();
 
     static {
         try {
-            // TODO: 2023/12/10 should scan from root package 
-            for (Class<?> cls : scanAndGetAllClasses("io.agistep.event")) {
+            // TODO: 2023/12/10 should scan from root package
+            for (Class<?> cls : ScanClassProvider.scanAllClassesIn("io.agistep")) {
                 for (Method method : cls.getDeclaredMethods()) {
                     if (method.isAnnotationPresent(EventHandler.class)) {
                         var handler = method.getAnnotation(EventHandler.class);
                         var handlerName = handler.payload().getName();
-                        handlers.put(handlerName, initHandler(cls.newInstance()));
+                        handlers.put(handlerName, initHandler(cls.getDeclaredConstructor().newInstance()));
                     }
                 }
             }
         } catch (Exception e) {
             throw new RuntimeException("Failed to initialize EventHandlerRegistry", e);
         }
-    }
-
-    public static Set<Class<?>> scanAndGetAllClasses(String packageName) throws IOException, ClassNotFoundException {
-        ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
-        assert classLoader != null;
-        String path = packageName.replace('.', '/');
-        Enumeration<URL> resources = classLoader.getResources(path);
-        Set<File> dirs = new HashSet<>();
-        while (resources.hasMoreElements()) {
-            URL resource = resources.nextElement();
-            dirs.add(new File(resource.getFile()));
-        }
-        Set<Class<?>> classes = new HashSet<>();
-        for (File directory : dirs) {
-            classes.addAll(findClasses(directory, packageName));
-        }
-        return classes;
-    }
-
-    private static Set<Class<?>> findClasses(File directory, String packageName) throws ClassNotFoundException {
-        Set<Class<?>> classes = new HashSet<>();
-        if (!directory.exists()) {
-            return classes;
-        }
-        File[] files = directory.listFiles();
-        for (File file : files) {
-            if (file.isDirectory()) {
-                assert !file.getName().contains(".");
-                classes.addAll(findClasses(file, packageName + "." + file.getName()));
-            } else if (file.getName().endsWith(".class")) {
-                classes.add(Class.forName(packageName + '.' + file.getName().substring(0, file.getName().length() - 6)));
-            }
-        }
-        return classes;
     }
 
     private static HandlerAdapter initHandler(Object aggregate) {
